refactor(agent): restructure testcase-subgraph to testcaseGeneration
- Rename directory from testcase-subgraph to testcaseGeneration for clarity
- Change graph.ts to index.ts with export name testcaseGeneration
- Rename invokeSaveToolNode to saveToolNode for consistency
- Remove unnecessary nesting (nodes/, routing/ directories)
- Update all import paths and references
- Remove unused zod dependency from agent package

This follows the directory structure refactoring plan in issue #5495
