+++ frontend/internal-packages/agent/package.json
@@ -24,8 +24,7 @@
     "langsmith": "0.3.61",
     "neverthrow": "8.2.0",
     "uuid": "11.1.0",
-    "valibot": "1.1.0",
-    "zod": "3.25.76"
+    "valibot": "1.1.0"
   },
   "devDependencies": {
     "@biomejs/biome": "2.1.4",
